Support `i128` and `u128`

+++ frb_example/pure_dart/lib/src/rust/api/dart_code.dart
@@ -13,7 +13,7 @@ class OpaqueStructWithDartCodeTwinNormal extends RustOpaque {
       : super.dcoDecode(wire, _kStaticData);
 
   OpaqueStructWithDartCodeTwinNormal.sseDecode(
-      int ptr, int externalSizeOnNative)
+      BigInt ptr, int externalSizeOnNative)
       : super.sseDecode(ptr, externalSizeOnNative, _kStaticData);
 
   static final _kStaticData = RustArcStaticData(
